requirements.txt
requests
python-telegram-bot
logging
subprocess
curses
asyncio